+++ src/migrations/20241004030722-create-client.js
@@ -8,7 +8,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       },
       dni: {
         type: Sequelize.STRING,
@@ -20,7 +20,7 @@ module.exports = {
         allowNull: false
       },
       user: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         allowNull: false,
         references: {
           model: 'User',
